requirements.txt
python-binance>=1.0.17
pandas>=2.0.0
numpy>=1.24.0
rich>=13.0.0
requests>=2.28.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
websocket-client>=1.6.0
# PyTorch (CPU). Install separately:
#   pip install torch --index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
